+++ src/Symfony/Component/Console/Formatter/OutputFormatter.php
@@ -12,7 +12,6 @@
 namespace Symfony\Component\Console\Formatter;
 
 use Symfony\Component\Console\Exception\InvalidArgumentException;
-use Symfony\Component\Console\Helper\OutputWrapperInterface;
 
 /**
  * Formatter class for console output.
@@ -145,8 +144,8 @@ class OutputFormatter implements WrappableOutputFormatterInterface
 
         $offset = 0;
         $output = '';
-        $openTagRegex = OutputWrapperInterface::TAG_OPEN_REGEX_SEGMENT;
-        $closeTagRegex = OutputWrapperInterface::TAG_CLOSE_REGEX_SEGMENT;
+        $openTagRegex = '[a-z](?:[^\\\\<>]*+ | \\\\.)*';
+        $closeTagRegex = '[a-z][^<>]*+';
         $currentLineLength = 0;
         preg_match_all("#<(($openTagRegex) | /($closeTagRegex)?)>#ix", $message, $matches, \PREG_OFFSET_CAPTURE);
         foreach ($matches[0] as $i => $match) {
@@ -162,7 +161,7 @@ class OutputFormatter implements WrappableOutputFormatterInterface
             $offset = $pos + \strlen($text);
 
             // opening tag?
-            if ($open = ('/' !== $text[1])) {
+            if ($open = '/' !== $text[1]) {
                 $tag = $matches[1][$i][0];
             } else {
                 $tag = $matches[3][$i][0] ?? '';