+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -12,9 +12,7 @@ CHANGELOG
 
  * Add support to display table vertically when calling setVertical()
  * Add method `__toString()` to `InputInterface`
- * Added `OutputWrapperInterface` and `OutputWrapper` to allow modifying your
-   wrapping strategy in `SymfonyStyle` or in other `OutputStyle`. Eg: you can
-   switch off to wrap URLs.
+ * Added `OutputWrapper` to prevent truncated URL in `SymfonyStyle::createBlock`.
  * Deprecate `Command::$defaultName` and `Command::$defaultDescription`, use the `AsCommand` attribute instead
  * Add suggested values for arguments and options in input definition, for input completion
  * Add `$resumeAt` parameter to `ProgressBar#start()`, so that one can easily 'resume' progress on longer tasks, and still get accurate `getEstimate()` and `getRemaining()` results.